Added map command. Refactor Bot
