+++ scripts/config.example.json
@@ -2,11 +2,5 @@
     "app_id": "",
     "app_secret": "",
     "guild_id": "",
-    "bot_token": "",
-    "events": [
-        {
-            "name": "High Five",
-            "type": 2
-        }
-    ]
+    "bot_token": ""
 }
